# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IContactMessage": {
				"properties": {
					"messageId": {
						"type": "number",
						"format": "double",
						"description": "Identifiant unique du message"
					},
					"email": {
						"type": "string",
						"description": "L'adresse e-mail"
					},
					"familyName": {
						"type": "string",
						"description": "Le nom de famille"
					},
					"givenName": {
						"type": "string",
						"description": "Le prénom"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female",
							"other",
							"unknown"
						],
						"description": "Le genre de l'utilisateur"
					},
					"conditions": {
						"type": "boolean",
						"description": "Boolean qui indique si l'utilisateur a accepté des conditions générales d'utilisation"
					},
					"newsletter": {
						"type": "boolean",
						"description": "Boolean qui indique si l'utilisateur a accepté de recevoir des communications"
					},
					"message": {
						"type": "string",
						"description": "Le text du message"
					}
				},
				"required": [
					"messageId",
					"email",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IContactMessage_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double",
						"description": "Le numéro de la page récupérée"
					},
					"limit": {
						"type": "number",
						"format": "double",
						"description": "Le nombre de lignes dans cette page"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "La quantité de lignes en total"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IContactMessage"
						},
						"type": "array",
						"description": "Les données retournées"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Identifiant de la ligne créée"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IContactMessage.Exclude_keyofIContactMessage.messageId__": {
				"properties": {
					"email": {
						"type": "string",
						"description": "L'adresse e-mail"
					},
					"familyName": {
						"type": "string",
						"description": "Le nom de famille"
					},
					"givenName": {
						"type": "string",
						"description": "Le prénom"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female",
							"other",
							"unknown"
						],
						"description": "Le genre de l'utilisateur"
					},
					"conditions": {
						"type": "boolean",
						"description": "Boolean qui indique si l'utilisateur a accepté des conditions générales d'utilisation"
					},
					"newsletter": {
						"type": "boolean",
						"description": "Boolean qui indique si l'utilisateur a accepté de recevoir des communications"
					},
					"message": {
						"type": "string",
						"description": "Le text du message"
					}
				},
				"required": [
					"email",
					"message"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IContactMessage.messageId_": {
				"$ref": "#/components/schemas/Pick_IContactMessage.Exclude_keyofIContactMessage.messageId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IContactMessageCreate": {
				"$ref": "#/components/schemas/Omit_IContactMessage.messageId_"
			},
			"IUpdateResponse": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IContactMessageCreate_": {
				"properties": {
					"email": {
						"type": "string",
						"description": "L'adresse e-mail"
					},
					"familyName": {
						"type": "string",
						"description": "Le nom de famille"
					},
					"givenName": {
						"type": "string",
						"description": "Le prénom"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female",
							"other",
							"unknown"
						],
						"description": "Le genre de l'utilisateur"
					},
					"conditions": {
						"type": "boolean",
						"description": "Boolean qui indique si l'utilisateur a accepté des conditions générales d'utilisation"
					},
					"newsletter": {
						"type": "boolean",
						"description": "Boolean qui indique si l'utilisateur a accepté de recevoir des communications"
					},
					"message": {
						"type": "string",
						"description": "Le text du message"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IContactMessageUpdate": {
				"$ref": "#/components/schemas/Partial_IContactMessageCreate_"
			},
			"IPhoto": {
				"properties": {
					"photoId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					}
				},
				"required": [
					"photoId",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IIndexResponse_IPhoto_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double",
						"description": "Le numéro de la page récupérée"
					},
					"limit": {
						"type": "number",
						"format": "double",
						"description": "Le nombre de lignes dans cette page"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "La quantité de lignes en total"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IPhoto"
						},
						"type": "array",
						"description": "Les données retournées"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-js-vanilla",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/contact": {
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IContactMessage_"
								}
							}
						}
					}
				},
				"description": "Récupérer des messages de contact",
				"security": [],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "CreateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouveau message",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IContactMessageCreate"
							}
						}
					}
				}
			}
		},
		"/contact/{messageId}": {
			"get": {
				"operationId": "ReadMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IContactMessage"
								}
							}
						}
					}
				},
				"description": "Récupérer un message à partir de son `messageId`",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un message de contact étant donnée son `messageId`",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IContactMessageUpdate"
							}
						}
					}
				}
			}
		},
		"/photo": {
			"get": {
				"operationId": "GetPhotos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IPhoto_"
								}
							}
						}
					}
				},
				"description": "Récupérer la liste de photos disponibled",
				"security": [],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}